Python-App-Server/requirements.txt
# flask-server/requirements.txt
Flask
Flask-SQLAlchemy
python-dotenv
psycopg2-binary
boto3
Flask-CORS
python-jose
requests
Flask-Migrate
